add denoise process
