+++ RustApp/Cargo.toml
@@ -87,6 +87,7 @@ rubato = "0.16.1"
 once_cell = "1.21.1"
 fix-path-env = { git = "https://github.com/tauri-apps/fix-path-env-rs" }
 env_logger = "0.11.7"
+nnnoiseless = "0.5.1"
 
 [build-dependencies]
 prost-build = "0.13"
